chore: relocate unit test
add new json snapshot serializer
