+++ package.json
@@ -28,6 +28,7 @@
     "emotion": "9.2.12",
     "enzyme": "3.9.0",
     "enzyme-adapter-react-16": "1.10.0",
+    "enzyme-to-json": "3.3.5",
     "eslint": "5.14.1",
     "eslint-config-google": "0.12.0",
     "eslint-config-prettier": "4.1.0",
