--- test/unit/__helper.js
@@ -1,5 +1,0 @@
-import path from 'path';
-
-export const parseConfigurationFile = (name) => {
-  return path.join(__dirname, `./partials/config/yaml/${name}.yaml`);
-};
