+++ test/unit/components/__mocks__/token.js
