+++ test/unit/components/copyToClipBoard.spec.js
@@ -6,8 +6,8 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import CopyToClipBoard from '../../../../src/webui/components/CopyToClipBoard';
-import { CopyIcon } from '../../../../src/webui/components/CopyToClipBoard/styles';
+import CopyToClipBoard from '../../../src/webui/components/CopyToClipBoard/index';
+import { CopyIcon } from '../../../src/webui/components/CopyToClipBoard/styles';
 
 describe('<CopyToClipBoard /> component', () => {
   let wrapper;
