+++ test/unit/components/package.spec.js
@@ -4,9 +4,9 @@
 
 import React from 'react';
 import { shallow } from 'enzyme';
-import Package from '../../../../src/webui/components/Package/index';
-import Tag from '../../../../src/webui/components/Tag/index';
-import { Version, WrapperLink, Field, OverviewItem } from '../../../../src/webui/components/Package/styles';
+import Package from '../../../src/webui/components/Package/index';
+import Tag from '../../../src/webui/components/Tag/index';
+import { Version, WrapperLink, Field, OverviewItem } from '../../../src/webui/components/Package/styles';
 
 
 /**
