+++ test/unit/components/packagelist.spec.js
@@ -4,8 +4,8 @@
 
 import React from 'react';
 import { mount } from 'enzyme';
-import PackageList from '../../../../src/webui/components/PackageList/index';
-import Help from '../../../../src/webui/components/Help/index';
+import PackageList from '../../../src/webui/components/PackageList/index';
+import Help from '../../../src/webui/components/Help/index';
 import { BrowserRouter } from 'react-router-dom';
 
 describe('<PackageList /> component', () => {
