+++ test/unit/components/search.spec.js
@@ -6,11 +6,11 @@
 import React from 'react';
 import { mount } from 'enzyme';
 
-import { Search } from '../../../../src/webui/components/Search/index';
+import { Search } from '../../../src/webui/components/Search/index';
 
-const SEARCH_FILE_PATH = '../../../../src/webui/components/Search/index';
-const API_FILE_PATH = '../../../../src/webui/utils/api';
-const URL_FILE_PATH = '../../../../src/webui/utils/url';
+const SEARCH_FILE_PATH = '../../../src/webui/components/Search/index';
+const API_FILE_PATH = '../../../src/webui/utils/api';
+const URL_FILE_PATH = '../../../src/webui/utils/url';
 
 // Global mocks
 const event = {
