+++ test/unit/components/store/logo.js
