+++ test/unit/components/store/packageMeta.js
